Correct catch nest level inside multi-peer try/catch blocks.
When a try/catch block has multiple catch elements, each of the catch blocks
is a 'peer' of the others, exceptions raised inside any of them will not be
handled by the peer catch. To do this, the compiler generates unwind code to
set the exception stack to the right position before jumping into the
exception handler function.

This means that while compiling the exception handler function, the catch
stack must match that from before entering the top-most catch peer. This was
done by not pushing non-local objects on the non local stack while recursing
to the peer catch block. However, aside from the nested catch blocks, the
nested statement *must* have the correct number of non-local goto blocks.

Instead of attempting to work around a shortage of non-local goto blocks
while compiling the statement inside the try/catch, the non-local stack is
stripped of peer catch blocks while compiling the exception handler.