Uninitialized value visible during MemCollect in Catch record.
Newly allocated memory is uninitialized and may contain garbage. If the
allocator may be called during initialization of the object, all object
fields must be initialized to a valid value before any further allocation
occurs. This wasn't happening in the Catch object as the exception member
was left unitialized while the continuation member was allocated.