Manage file buffer chains with explicit malloc/free.
Buffer chains must have life equal to the file they belong to; an
unreferenced file may still have buffered data, so the buffers cannot be
reclaimed. The nickle GC does not have a separate pass that checks to see
which objects can be freed, rather it assumes that such objects are
self-contained.

Hence, anything hanging off of an object which can refuse to
be reclaimed must not have been allocated from the nickle GC system. Yes,
this is a significant limitation to this allocator.

Someday I'll figure out a good general fix for this problem, right now we'll
stop segfaults when applications drop file references without closing them.