Add signal exceptions, treat SIGINT as raise signal(2)
SIGINT used to interrupt the current thread and enter the debugger with the
thread 'interrupted'; in this state, the thread could be continued. However,
this means that threads cannot catch SIGINT and cleanup at process exit.
This commit changes all this around so that SIGINT simply raises a signal
exception in all threads.

This patch also keeps the interpreter running until all threads have exited.