Allow background nickle to not poll on tty ownership
When stdin is connected to a terminal but nickle is not the foreground
process on that terminal, the io code would poll to wait for ownership to
flip back to nickle so that reads could be performed without generating a
signal. Now, nickle waits until someone actually tries to read from the
terminal before starting to poll. This means that simple background
processing nickle programs will not poll every 100ms.