Handle OpFarJump in CompileReachable
A FarJump within a catch block references instructions one or more
frames outside of the instruction context. When checking for reachable
code, look down inside the catch blocks to see if any of the FarJumps
within them touch the target instruction.

Fixes this example:

void foo() {
	for (;;)
		try {
		} catch uninitialized_value(string x) {
			break;
		}
}

Without this fix, the 'break' will not get noticed and no ReturnVoid
will be appended to the object code for 'foo', leaving the break
dangling in space.

Signed-off-by: Keith Packard <keithp@keithp.com>