Open quote character is a single char, not a string
When checking to see if a string needs to be dequoted, check the first
character of the string against the quote character, rather than
checking the whole string.

Probably using strings for the quote markers wasn't the best API
choice as this would have been caught by typechecking had the quote
characters been an int instead.

Signed-off-by: Keith Packard <keithp@keithp.com>