Add Math::e_value. Fix exp and log precision error.
e_value computes e to any desired precision.

exp was incorrectly truncating the value just before the last
computation step, which caused it to lose significant precision.

log was range-reducing to 0-1 instead of 2-3, which put it outside the
range of rapid convergence. It also needed another newtons method
iteration step.

Signed-off-by: Keith Packard <keithp@keithp.com>