Remove incorrect tree walking in CompileSizeDimensions
CompileSizeDimensions was incorrectly interpreting the tree structure
for bare value nodes -- it is either passed an ARRAY, or a single
expression value, which can either be a comprehension, an empty
initializer or a simple expression. The simple expression case was
incorrectly looking for a '...' element to the left, which cannot
happen. It further recursed if it hadn't reached the end of the
dimensions, which is pointless as that case will generate an
error when the initializers are compiled later.

Debian-bug: #911926
Signed-off-by: Keith Packard <keithp@keithp.com>