Allow repeat initializers in implicit sized arrays for small dimensions
When initializing an implicitly sized multi-dimensional array, permit
repeat initializers ("...") to be used to fill out smaller dimensions.
This is done by requiring that the maximum number of elements in each
dimension be taken from an initalizer without a repeat element.

This also requires allowing repeat initializers to be useful for
resizable arrays, which required a change to the execution machinery
as well.

Signed-off-by: Keith Packard <keithp@keithp.com>